Namespaces refactor
