+++ src/BbsSignatures.Tests/BbsBlindSignTests.cs
@@ -1,8 +1,4 @@
-﻿using BbsSignatures.Bls;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
+﻿using System.Linq;
 using System.Threading.Tasks;
 using Xunit;
