+++ src/BbsSignatures/BbsProvider.cs
@@ -1,5 +1,4 @@
-﻿using BbsSignatures.Bls;
-using System;
+﻿using System;
 using System.Collections.ObjectModel;
 using System.Linq;
 using System.Threading.Tasks;
