+++ src/BbsSignatures/Keys/BlsSecretKey.cs
@@ -1,9 +1,6 @@
-﻿using BbsSignatures.Bls;
-using System;
+﻿using System;
 using System.Collections.ObjectModel;
 using System.Linq;
-using System.Runtime.InteropServices;
-using System.Runtime.InteropServices.ComTypes;
 using System.Text;
 
 namespace BbsSignatures
