+++ src/BbsSignatures/Models/SignatureProofStatus.cs
@@ -1,4 +1,4 @@
-﻿namespace BbsSignatures.Bls
+﻿namespace BbsSignatures
 {
     public enum SignatureProofStatus
     {
