+++ src/BbsSignatures/UnmanagedMemoryContext.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Runtime.InteropServices;
-using BbsSignatures.Bls;
 
 namespace BbsSignatures
 {
